SM-1287: Refactor all secret commands out of main.rs
