+++ crates/bws/src/command/mod.rs
@@ -1,4 +1,5 @@
 pub(crate) mod project;
+pub(crate) mod secret;
 
 use std::{path::PathBuf, str::FromStr};
